Moved project to new VS Studio project format Added a nuget/TpRdlcViewer.2.0.0.nupkg package file
