+++ RDLCReport/RDLCReport.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RDLCDemo", "RDLCDemo\RDLCDemo.csproj", "{D81CA664-CBF9-4EB7-9E63-203A8B47850B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TimePunch.Rdlc", "MrDSoft.RDLC\TimePunch.Rdlc.csproj", "{5578EECF-3252-44EF-897A-C991153EF2AE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TimePunch.Rdlc", "TimePunch.Rdlc\TimePunch.Rdlc.csproj", "{5A659487-24D3-4C59-9122-00211F8B436D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{D81CA664-CBF9-4EB7-9E63-203A8B47850B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D81CA664-CBF9-4EB7-9E63-203A8B47850B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D81CA664-CBF9-4EB7-9E63-203A8B47850B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5578EECF-3252-44EF-897A-C991153EF2AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5578EECF-3252-44EF-897A-C991153EF2AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5578EECF-3252-44EF-897A-C991153EF2AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5578EECF-3252-44EF-897A-C991153EF2AE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5A659487-24D3-4C59-9122-00211F8B436D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5A659487-24D3-4C59-9122-00211F8B436D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5A659487-24D3-4C59-9122-00211F8B436D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5A659487-24D3-4C59-9122-00211F8B436D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
