+++ RDLCReport/TimePunch.Rdlc/RdlcPrinterDialog.xaml
@@ -1,0 +1,156 @@
+﻿<controls:ContentDialog
+    x:Class="TimePunch.Rdlc.RdlcPrinterDialog"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:controls="http://schemas.modernwpf.com/2019"
+    xmlns:resources="clr-namespace:TimePunch.Rdlc.Resources"
+    PrimaryButtonClick="OK_Click"
+    PrimaryButtonText="{x:Static resources:ViewerRes.Print}"
+    IsPrimaryButtonEnabled="True"
+    CloseButtonCommand="Close"
+    CloseButtonText="{x:Static resources:ViewerRes.Close}"
+    Loaded="Window_Loaded">
+
+    <!--
+        CREDIT : 2013-2018 Derek Tremblay (abbaye), 2013 Martin Savard
+        https://github.com/abbaye/RDLCPrinter
+    -->
+
+    <Grid>
+
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*"></ColumnDefinition>
+            <ColumnDefinition Width="*"></ColumnDefinition>
+        </Grid.ColumnDefinitions>
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+
+        </Grid.RowDefinitions>
+
+        <!-- Printer Settings -->
+        <Label
+            Grid.Row="0" Grid.Column="0"
+            Content="{x:Static resources:ViewerRes.PrinterSettings}"
+            Margin="6"
+            FontSize="20"/>
+
+        <!-- Printer name -->
+        <Label
+            Grid.Row="1" Grid.Column="0"
+            VerticalAlignment="Center"
+            Margin="6"
+            Content="{x:Static resources:ViewerRes.PrinterName}" />
+
+        <ComboBox
+            Grid.Row="1" Grid.Column="1"
+            HorizontalAlignment="Stretch"
+            Margin="6"
+            x:Name="PrinterName"
+            SelectionChanged="CboImprimanetNom_SelectionChanged" />
+
+        <!-- Status -->
+        <Label
+            Grid.Row="2" Grid.Column="0"
+            VerticalAlignment="Center"
+            Margin="6"
+            Content="{x:Static resources:ViewerRes.PrinterState}" />
+
+        <Label
+            x:Name="PrinterState"
+            Grid.Row="2" Grid.Column="1"
+            VerticalAlignment="Center"
+            Margin="6"
+            Content=""/>
+
+        <!-- Copy Settings -->
+        <Label
+            Grid.Row="5" Grid.Column="0"
+            Content="{x:Static resources:ViewerRes.CopySettings}"
+            Margin="6"
+            FontSize="20"/>
+
+        <!-- Number of copies -->
+        <Label
+            Grid.Row="6" Grid.Column="0"
+            VerticalAlignment="Center"
+            Margin="6"
+            Content="{x:Static resources:ViewerRes.NumberOfCopies}"/>
+
+        <controls:NumberBox
+            Grid.Row="6" Grid.Column="1"
+            x:Name="NumberOfCopySpinner"
+            Value="1"
+            SpinButtonPlacementMode="Inline"
+            Minimum="1"
+            Maximum="500"
+            SmallChange="1"
+            LargeChange="10" 
+            Margin="6"/>
+
+        <!-- Page Range Settings -->
+        <Label
+            Grid.Row="7" Grid.Column="0"
+            Content="{x:Static resources:ViewerRes.PageRange}"
+            Margin="6"
+            FontSize="20"/>
+
+        <StackPanel Orientation="Vertical" Grid.Row="8" Grid.Column="0">
+            <RadioButton
+                x:Name="AllPages"
+                VerticalAlignment="Center"
+                IsChecked="True"
+                Margin="6"
+                Content="{x:Static resources:ViewerRes.PrintAll}"/>
+            <RadioButton
+                x:Name="PageRange"
+                VerticalAlignment="Center"
+                Margin="6"
+                Content="{x:Static resources:ViewerRes.PrintPageRange}"/>
+        </StackPanel>
+
+        <!-- First page -->
+        <StackPanel Orientation="Horizontal" x:Name="PageChoiceStackPanel" Grid.Row="9" Grid.Column="0" Grid.ColumnSpan="2"
+                    IsEnabled="{Binding ElementName=PageRange, Path=IsChecked}">
+
+            <Label
+                Content="{x:Static resources:ViewerRes.FromPage}"
+                VerticalAlignment="Center"
+                Margin="6"/>
+
+            <controls:NumberBox
+                x:Name="FirstPageSpinner"
+                Value="1"
+                SpinButtonPlacementMode="Hidden"
+                SmallChange="1"
+                LargeChange="10" 
+                Margin="6"/>
+
+            <Label
+                Content="{x:Static resources:ViewerRes.ToPage}"
+                VerticalAlignment="Center"
+                Margin="6"/>
+
+            <!-- Last page -->
+            <controls:NumberBox
+                x:Name="LastPageSpinner"
+                Value="10"
+                SpinButtonPlacementMode="Hidden"
+                SmallChange="1"
+                LargeChange="10" 
+                Margin="6"/>
+        </StackPanel>
+    </Grid>
+
+</controls:ContentDialog>
