+++ RDLCReport/TimePunch.Rdlc/TimePunch.Rdlc.csproj
@@ -1,0 +1,81 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net472</TargetFramework>
+    <Version>2.0.0</Version>
+    <LangVersion>latest</LangVersion>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <AssemblyName>TpRdlcViewer</AssemblyName>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <Authors>Gerhard Stephan</Authors>
+    <Company>TimePunch KG</Company>
+    <Description>WPF Report Viewer for RDLC files.</Description>
+    <PackageProjectUrl></PackageProjectUrl>
+    <PackageIcon>Icon.png</PackageIcon>
+    <RepositoryUrl>https://github.com/BitKFu/RDLCPrinter</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageTags>wpf; rdlc; viewer; modern</PackageTags>
+    <SignAssembly>True</SignAssembly>
+    <AssemblyOriginatorKeyFile>TpRdlcViewer.snk</AssemblyOriginatorKeyFile>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove="RdlcPrinterDialog.xaml" />
+    <None Remove="RDLCReportViewer.xaml" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="..\..\Icon.png">
+      <Pack>True</Pack>
+      <PackagePath>\</PackagePath>
+    </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.ReportingServices.ReportViewerControl.Winforms" Version="150.1484.0" />
+    <PackageReference Include="Microsoft.SqlServer.Types" Version="14.0.1016.290" />
+    <PackageReference Include="ModernWpfUI.Signed" Version="0.9.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Page Include="RdlcPrinterDialog.xaml">
+      <Generator>XamlIntelliSenseFileGenerator</Generator>
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Page>
+    <Page Include="RdlcReportViewer.xaml">
+      <Generator>XamlIntelliSenseFileGenerator</Generator>
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Page>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
+    <Reference Include="System.Printing" />
+    <Reference Include="System.Windows.Forms" />
+    <Reference Include="System.Xaml" />
+    <Reference Include="WindowsBase" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Update="RdlcPrinterDialog.xaml.cs">
+      <DependentUpon>RdlcPrinterDialog.xaml</DependentUpon>
+    </Compile>
+    <Compile Update="RDLCReportViewer.xaml.cs">
+      <DependentUpon>RdlcReportViewer.xaml</DependentUpon>
+    </Compile>
+    <Compile Update="Resources\ViewerRes.Designer.cs">
+      <DependentUpon>ViewerRes.resx</DependentUpon>
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+    </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Update="Resources\ViewerRes.resx">
+      <LastGenOutput>ViewerRes.Designer.cs</LastGenOutput>
+      <Generator>PublicResXFileCodeGenerator</Generator>
+    </EmbeddedResource>
+  </ItemGroup>
+
+</Project>
